+++ src/matomo.ts
@@ -1,0 +1,26 @@
+/** @file Adaptations of the vue3-matomo plugin */
+import { inject, toRef, type App } from "vue";
+import { createVueMatomo, matomoKey } from "vue3-matomo";
+import router from "./router/router";
+
+/** Vue plugin that wraps the original Matomo plugin but only enables it if configured */
+export const matomo = {
+  install(app: App) {
+    const { VITE_MATOMO_URL, VITE_MATOMO_ID } = import.meta.env;
+    if (VITE_MATOMO_URL && VITE_MATOMO_ID) {
+      app.use(
+        createVueMatomo({
+          // Url expected without trailing slash
+          host: VITE_MATOMO_URL.replace(/\/$/, ""),
+          siteId: VITE_MATOMO_ID,
+          router,
+        }),
+      );
+    }
+  },
+};
+
+/** Replacement for `useMatomo()` that works even if the plugin is not installed */
+export function useMatomo() {
+  return toRef(inject(matomoKey, undefined));
+}